+++ Box.V2/Managers/BoxMetadataManager.cs
@@ -5,6 +5,7 @@ using Box.V2.Exceptions;
 using Box.V2.Extensions;
 using Box.V2.Models;
 using Box.V2.Services;
+using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -314,7 +315,79 @@ namespace Box.V2.Managers
             return response.ResponseObject;
         }
 
+        /// <summary>
+        /// Allows you to query by metadata on Box items
+        /// </summary>
+        /// <param name="from">The template used in the query. Must be in the form scope.templateKey</param>
+        /// <param name="ancestorFolderId">The folder_id to which to restrain the query</param>
+        /// <param name="query">The logical expression of the query</param>
+        /// <param name="queryParameters">Required if query present. The arguments for the query</param>
+        /// <param name="indexName">The name of the Index to use</param>
+        /// <param name="orderBy">A list of BoxMetadataQueryOrderBy objects that contain field_key(s) to order on and the corresponding direction(s)</param>
+        /// <param name="limit">The maximum number of items to return in a page. The default is 100 and the max is 1000.</param>
+        /// <param name="marker">The marker to use for requesting the next page</param>
+        /// <param name="autoPaginate">Whether or not to auto-paginate to fetch all items; defaults to false.</param>
+        /// <returns>A collection of items and their associated metadata</returns>
+        public async Task<BoxCollectionMarkerBased<BoxMetadataQueryItem>> ExecuteMetadataQueryAsync(string from, string ancestorFolderId, string query = null, Dictionary<string, object> queryParameters = null, string indexName = null, List<BoxMetadataQueryOrderBy> orderBy = null, int limit = 100, string marker = null, bool autoPaginate = false)
+        {
+            from.ThrowIfNullOrWhiteSpace("from");
+            ancestorFolderId.ThrowIfNullOrWhiteSpace("ancestorFolderId");
 
+            dynamic queryObject = new JObject();
+
+            queryObject.from = from;
+            queryObject.ancestor_folder_id = ancestorFolderId;
+            queryObject.limit = limit;
+
+            if (query != null)
+            {
+                queryObject.query = query;
+            }
+
+            if (queryParameters != null)
+            {
+                queryObject.query_params = _converter.Serialize(queryParameters);
+            }
+
+            if (indexName != null)
+            {
+                queryObject.use_index = indexName;
+            }
+
+            if (orderBy != null)
+            {
+                List<JObject> orderByList = new List<JObject>();
+                foreach (var order in orderBy)
+                {
+                    dynamic orderByObject = new JObject();
+                    orderByObject.field_key = order.FieldKey;
+                    orderByObject.direction = order.Direction.ToString();
+                    orderByList.Add(orderByObject);
+                }
+                queryObject.order_by = _converter.Serialize(orderByList); ;
+            }
+
+            if (marker != null)
+            {
+                queryObject.marker = marker;
+            }
+
+            string queryStr = queryObject.ToString();
+
+            BoxRequest request = new BoxRequest(_config.MetadataQueryUri)
+                .Method(RequestMethod.Post)
+                .Payload(queryStr);
+
+            if (autoPaginate)
+            {
+                return await AutoPaginateMarkerMetadataQuery<BoxMetadataQueryItem>(request, limit);
+            }
+            else
+            {
+                IBoxResponse<BoxCollectionMarkerBased<BoxMetadataQueryItem>> response = await ToResponseAsync<BoxCollectionMarkerBased<BoxMetadataQueryItem>>(request).ConfigureAwait(false);
+                return response.ResponseObject;
+            }
+        }
 
         //************************************
         //Private methods