+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 - Updated retry logic to retry on 503 status codes returned by the API
+- Added the ability to query Box items based on their metadata. The method to do so is `MetadataManager.ExecuteMetadataQueryAsync()`.
+- Added `TrashedAt` field to `BoxItem` objects (file, folder, weblink).
 
 ## 3.21.0 [2019-12-05]
 - Added `fields` parameter to `UsersManager.GetUserInformationAsync()`